fixed most syntax issues, some still persist, changed most tests for new syntax, some are still left
