+++ src/general/types.jl
@@ -123,16 +123,6 @@ Regular updates with no blocking
 struct NoBlocking <: BlockingSchedule end
 
 
-
-abstract type ActionType end
-
-struct Verbose <: ActionType end
-struct SavePath <: ActionType end
-struct Imputation <: ActionType end
-struct ParamUpdate <: ActionType end
-struct Readjust <: ActionType end
-
-
 #TODO implement Jeffrey's priors
 """
     ImproperPrior
@@ -140,10 +130,10 @@ struct Readjust <: ActionType end
 Flat prior
 """
 struct ImproperPrior end
-logpdf(::ImproperPrior, θ) = 0.0
+logpdf(::ImproperPrior, coord_idx, θ) = 0.0
 
-struct ImproperPosPrior{T} end
-logpdf(::ImproperPosPrior{Val{T}}, θ) where T = -log(θ[T])
+struct ImproperPosPrior end
+logpdf(::ImproperPosPrior, coord_idx, θ) = -sum(log.(θ[indices(coord_idx)]))
 
 valtype(::Val{T}) where T = T
 
@@ -161,33 +151,6 @@ julia> idx(Val((true, false, false, true, true)))
 function idx(::Val{T}) where T
     tuple((i for i in 1:length(T) if T[i])...)
 end
-
-#NOTE already defined in coordinate_access
-"""
-    moveToProperPlace(ϑ, θ, ::Val{T}) where {T}
-
-Update parameter vector `θ` at indices specified by `Val{T}` with the
-values collected in `ϑ`.
-
-# Examples
-```julia-repl
-julia> moveToProperPlace([10,20,30], [1,2,3,4,5],
-                         Val((true, false, true, false, true)))
-5-element Array{Float64,1}:
- 10.0
-  2.0
- 20.0
-  4.0
- 30.0
-```
-"""
-function move_to_proper_place(ϑ, θ, ::Val{v}) where {v}
-    θᵒ = copy(θ)
-    idxNew = [i for i in 1:length(v) if v[i]]
-    θᵒ[idxNew] = ϑ
-    θᵒ
-end
-
 
 get_aux_flag(::Any) = nothing
