+++ src/mcmc/mcmc_components.jl
@@ -18,6 +18,12 @@ function reset!(at::AccptTracker{S}) where S
     at.prop = zero(S)
 end
 
+function reset!(at::Vector{<:AccptTracker})
+    for a in at
+        reset!(a)
+    end
+end
+
 function acceptance_rate(at::AccptTracker{S}) where S
     at.prop == zero(S) && return 0.0
     at.accpt/at.prop
@@ -65,21 +71,22 @@ Needs to have a prior, transition kernel, acceptance rate tracker, information
 about the which coordinates are being updated and an object with auxiliary,
 use-specific information say whether to recompute_ODEs
 """
-struct MCMCParamUpdate{UpdtType,S,T,U,V} <: MCMCUpdate
+struct ParamUpdate{UpdtType,S,T,U,V} <: MCMCUpdate
     updt_coord::S
     t_kernel::T
     priors::U
-    accpt_tracker::AccptTracker
+    accpt_tracker::AccptTracker{Int64}
     accpt_history::Vector{Bool}
     aux::V
     readjust_param::ReadjustT
 
 
-    function MCMCUpdate(::UpdtType, updt_coord, θ, t_kernel::T, priors::U,
-                        aux::V, readjust_param=(100, 0.1, -999, 999, 0.234, 50)
-                        ) where {UpdtType<:ParamUpdateType,T,U,V<:AuxiliaryInfo}
+    function ParamUpdate(::UpdtType, updt_coord, θ, t_kernel::T, priors,
+                         aux::V, readjust_param=(100, 0.1, 0.0, 999.9, 0.234, 50)
+                         ) where {UpdtType<:ParamUpdateType,T,V<:AuxiliaryInfo}
         updt_coord = reformat_updt_coord(updt_coord, θ)
-        S = typeof(updt_coord)
+        priors = Priors(priors, updt_coord)
+        S, U = typeof(updt_coord), typeof(priors)
         new{UpdtType,S,T,U,V}(updt_coord, t_kernel, priors, AccptTracker(0),
                               Bool[], aux, named_readjust(readjust_param))
     end
@@ -112,37 +119,38 @@ appropriate Val{(...)}() object is passed and nothing is done, use at your own r
 """
 reformat_updt_coord(updt_coord, θ) = updt_coord
 
-
-function readjust!(pu::MCMCParamUpdate, corr_mat, mcmc_iter)
+function readjust!(pu::ParamUpdate, corr_mat, mcmc_iter)
     at = pu.accpt_tracker
     p = pu.readjust_param
-    readjust!(pu.t_kernel, at, p, corr_mat, mcmc_iter)
+    readjust!(pu.t_kernel, at, p, corr_mat, mcmc_iter, pu.updt_coord)
     reset!(at)
 end
 
+readjust!(::Nothing, params...) = nothing
+
 aux_params(::Any, aux) = aux
 
-function register_accpt!(pu::MCMCParamUpdate, acc)
+function register_accpt!(pu::ParamUpdate, acc)
     register_accpt!(pu.accpt_tracker, acc)
     push!(accpt_history, acc)
 end
 #===============================================================================
                         Definition of an imputation step
 ===============================================================================#
-struct MCMCImputation{B,ST,T} <: MCMCUpdate
-    accpt_tracker::Vector{AccptTracker}
+struct Imputation{B,ST,T} <: MCMCUpdate
+    accpt_tracker::Vector{AccptTracker{Int64}}
     accpt_history::Vector{T}
     ρs::Vector{Float64}
     blocking::B
     solver::ST
     readjust_param::ReadjustT
 
-    function MCMCImputation(blocking::B, ρ::S, solver::ST=Ralston3(),
+    function Imputation(blocking::B, ρ::S, solver::ST=Ralston3(),
                             readjust_param=(100, 0.1, 0.00001, 0.99999, 0.234, 50)
                             ) where {B,S<:Number,ST<:ODESolverType}
         accpt_tracker = [AccptTracker(0) for _ in 1:length(blocking)]
-        accpt_history = NTuple{length(blocking),Bool}[]
-        T = typeof(accpt_history)
+        T = NTuple{length(blocking),Bool}
+        accpt_history = T[]
         readjust_param = named_readjust(readjust_param)
         ρs = fill(ρ, length(blocking))
         new{B,ST,T}(accpt_tracker, accpt_history, ρs, blocking, solver,
@@ -150,13 +158,13 @@ struct MCMCImputation{B,ST,T} <: MCMCUpdate
     end
 end
 
-function readjust!(pu::MCMCImputation, ::Any, mcmc_iter)
+function readjust!(pu::Imputation, ::Any, mcmc_iter)
     at = pu.accpt_tracker
     p = pu.readjust_param
     δ = compute_δ(p, mcmc_iter)
     ar = [acceptance_rate(tracker) for tracker in at]
     for i in 1:length(pu.ρs)
-        pu.ρs[i] = compute_ϵ(ws.ρ[i], p, ar[i], δ, -1.0, logit, sigmoid)
+        pu.ρs[i] = compute_ϵ(pu.ρs[i], p, ar[i], δ, -1.0, logit, sigmoid)
     end
     display_new_ρ(ar, pu.ρs)
     reset!(at)
@@ -170,9 +178,9 @@ function display_new_ρ(accpt_rates, ρs)
     print("\n")
 end
 
-aux_params(pu::MCMCImputation, aux) = pu.blocking
+aux_params(pu::Imputation, aux) = pu.blocking
 
-function register_accpt!(pu::MCMCImputation, acc)
+function register_accpt!(pu::Imputation, acc)
     register_accpt!(pu.accpt_tracker, acc)
     push!(pu.accpt_history, tuple(acc...))
 end
@@ -225,7 +233,7 @@ function Base.iterate(iter::MCMCSchedule, state=(iter.start, 1))
     actions = (idx=iter.updt_idx[element],
                save=(i > at.warm_up && i % at.save == 0),
                verbose=(i % at.verbose == 0),
-               param_updt=(at.param_updt && i > at.warm_up),
+               param_updt=(i > at.warm_up),
                readjust=at.readjust(i),
                fuse=at.fuse(i),
                iter=i)
@@ -240,46 +248,3 @@ function Base.eltype(iter::MCMCSchedule)
                Tuple{Array{Int64,1},Bool,Bool,Bool,Bool,Bool,Int64}}
 end
 transition(schedule::MCMCSchedule, elem) = mod1(elem + 1, length(schedule.updt_idx))
-
-
-
-#===============================================================================
-                                The main routine
-===============================================================================#
-
-function mcmc(setup_mcmc::MCMCSetup, schedule::MCMCSchedule, setup::T) where T <: ModelSetup
-    ws, ll, θ = create_workspace(setup)
-    ws_mcmc = create_workspace(setup_mcmc, schedule, θ)
-
-    aux = nothing
-    for step in schedule
-        step.save && save_imputed!(ws)
-        for i in step.idx
-            ws = next(ws, ws_mcmc.updates[i])
-            ll, acc, θ = update!(ws_mcmc.updates[i], ws, θ, ll, step, aux)
-            aux = aux_params(ws_mcmc.updates[i], aux)
-            update!(ws_mcmc, acc, θ, i)
-            step.verbose && print("\n")
-        end
-        step.verbose && print("-----------------------------------------------",
-                              "------\n")
-        step.readjust && readjust!(ws_mcmc)
-        step.fuse && fuse!(ws_mcmc, schedule)
-    end
-    display_summary(ws, ws_mcmc)
-    ws
-end
-
-
-
-
-
-foo = MCMCSchedule(10, [[1,2,3],[3,4],[5,6,7,8]], (1,2,3,x->(x%5==0),true, x->(x%4==0)))
-
-
-for (i,f) in enumerate(foo)
-    if i == 5
-        foo.updt_idx[1][1]=10000
-    end
-    print(f, "\n")
-end
